+++ lib/index.js
@@ -1,4 +1,4 @@
-import Attachment from "components/Attachment/index";
+import Attachments from "components/Attachments/index";
 import { isEditorEmpty } from "utils/common";
 
 import Editor from "./components/Editor";
@@ -7,4 +7,11 @@ import Menu from "./components/Editor/Menu";
 import EditorContent from "./components/EditorContent";
 import "./index.scss";
 
-export { Editor, EditorContent, Menu, isEditorEmpty, FormikEditor, Attachment };
+export {
+  Editor,
+  EditorContent,
+  Menu,
+  isEditorEmpty,
+  FormikEditor,
+  Attachments,
+};
